Sphere picking is now working

+++ OpenGL/MouseRay.cpp
@@ -5,8 +5,9 @@ MouseRay::MouseRay(void) {
   ray = glm::vec3(1.0f);
 }
 
-void MouseRay::Update(float xPos, float yPos, int width, int height,
+void MouseRay::Update(double xPos, double yPos, int width, int height,
 		      glm::mat4 pMat, glm::mat4 vMat, glm::vec3 orig) {
+
   float normx = (2.0f * xPos) / width - 1.0f;
   float normy = 1.0f - (2.0f * yPos) / height;
   glm::vec4 clipSpace = glm::vec4(normx, normy, -1.0, 1.0);
@@ -15,6 +16,7 @@ void MouseRay::Update(float xPos, float yPos, int width, int height,
   glm::vec4 worldSpace = glm::inverse(vMat) * eye;
   glm::vec3 space = glm::vec3(worldSpace.x, worldSpace.y, worldSpace.z);
   ray = glm::normalize(space);
+  
   origin = orig;
 }
